backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0

# Core dependencies
openai==0.28.1
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For better CLI experience
colorama>=0.4.6

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

